# docs/swagger.yaml
basePath: /
definitions:
  model.NewsListItem:
    properties:
      content:
        type: string
      date:
        type: string
      headline:
        type: string
      publisher:
        type: string
      time:
        type: string
    type: object
  model.NewsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.NewsListItem'
        type: array
      pagination:
        $ref: '#/definitions/model.Pagination'
    type: object
  model.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
info:
  contact:
    email: support@onelineai.com
    name: OLA B2B Team
  description: |-
    API server for Hana Securities translated news service.
    Provides access to Japanese (Minkabu) and Chinese (Wind) news translated to Korean.
  license:
    name: Proprietary
    url: https://onelineai.com
  title: Hana Securities News API
  version: "1.0"
paths:
  /health:
    get:
      description: Check service health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
  /v1/news/{country}:
    get:
      consumes:
      - application/json
      description: Get paginated list of translated news articles for a specific country
      parameters:
      - description: Country code (JP or CN)
        in: path
        name: country
        required: true
        type: string
      - description: Filter by ticker/stock code
        in: query
        name: ticker
        type: string
      - description: Start time (RFC3339 format)
        in: query
        name: from
        type: string
      - description: End time (RFC3339 format)
        in: query
        name: to
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.NewsListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List news by country
      tags:
      - news
  /v1/news/{country}/{exchange}:
    get:
      consumes:
      - application/json
      description: Get paginated list of translated news articles for a specific country
        and exchange
      parameters:
      - description: Country code (CN)
        in: path
        name: country
        required: true
        type: string
      - description: Exchange code (HK, SH, SZ, BJ)
        in: path
        name: exchange
        required: true
        type: string
      - description: Filter by ticker code (without exchange suffix)
        in: query
        name: ticker
        type: string
      - description: Start time (RFC3339 format)
        in: query
        name: from
        type: string
      - description: End time (RFC3339 format)
        in: query
        name: to
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.NewsListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List news by country and exchange
      tags:
      - news
schemes:
- http
- https
swagger: "2.0"